# docs/swagger.yaml
definitions:
  helper.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  model.Comment:
    properties:
      comment_message:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo:
        $ref: '#/definitions/model.Photo'
      photo_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.Photo:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photo_caption:
        example: photo caption
        type: string
      photo_title:
        example: photo title
        type: string
      photo_url:
        example: photo url
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
...
  model.SwagPhoto:
    properties:
      photo_caption:
        example: Your caption
        type: string
      photo_title:
        example: Photo Title
        type: string
      photo_url:
        example: http://img-url.com
        type: string
    type: object
...
  model.User:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      created_at:
        type: string
      id:
        type: integer
      photos:
        items:
          $ref: '#/definitions/model.Photo'
        type: array
      updated_at:
        type: string
      user_age:
        type: integer
      user_email:
        type: string
      user_password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /photos:
    get:
      description: Get all photos from specific user ID that was taken from the token
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Photo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Return all photos from this user
      tags:
      - Photo
    post:
      consumes:
      - application/json
      description: |-
        post a json data to upload photo
        Validates: (Photo Title, Photo URL)
      parameters:
      - description: photo
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/model.SwagPhoto'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Upload photo
      tags:
      - Photo
  /photos/{id}:
    delete:
      description: |-
        Return forbidden if the user is not the owner of the photo
        Validates: (Photo Title, Photo URL)
      parameters:
      - description: photo id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Delete certain photo that is specified by the id
      tags:
      - Photo
    get:
      consumes:
      - application/json
      description: Input valid photo id through parameter
      parameters:
      - description: photo id
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Photo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Get photo by photo id
      tags:
      - Photo
    put:
      consumes:
      - application/json
      description: |-
        Return forbidden if the user is not the owner of the photo
        Validates: (Photo Title, Photo URL)
      parameters:
      - description: photo id
        in: path
        name: id
        required: true
        type: integer
      - description: body
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/model.SwagPhoto'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helper.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helper.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.Response'
      security:
      - Bearer: []
      summary: Update certain photo that is specified by the id
      tags:
      - Photo
...
swagger: "2.0"